DP-691 Workaround for PGAdmin bug
  - Added logic to check for existing server configurations in PGAdmin's PostgreSQL database before adding new ones.

+++ terragrunt/tools/pgadmin/configs/custom-entrypoint.sh
@@ -2,7 +2,18 @@
 
 echo "Configuring pgAdmin..."
 
-cat <<EOF > /pgadmin4/servers.json
+export CONFIG_DATABASE_URI="postgresql://${PGADMIN_DATABASE_USERNAME}:${PGADMIN_DATABASE_PASSWORD}@${PGADMIN_DATABASE_HOST}:5432/${PGADMIN_DATABASE_NAME}"
+
+check_sirsi_exists() {
+    PGPASSWORD=$PGADMIN_DATABASE_PASSWORD psql -h $PGADMIN_DATABASE_HOST -U $PGADMIN_DATABASE_USERNAME -d $PGADMIN_DATABASE_NAME -tAc "SELECT 1 FROM public.server WHERE name = 'SIRSI'" | grep -q 1
+    return $?
+}
+
+if check_sirsi_exists; then
+    echo "SIRSI server already exists. Skipping adding all servers..."
+else
+    echo "Adding all servers..."
+    cat <<EOF > /pgadmin4/servers.json
 {
     "Servers": {
         "1": {
@@ -26,27 +37,24 @@ cat <<EOF > /pgadmin4/servers.json
         "3": {
             "Name": "PGAdmin",
             "Group": "Servers",
-            "Host": "${PGADMIN_CONFIG_DATABASE_HOST}",
+            "Host": "${PGADMIN_DATABASE_HOST}",
             "Port": 5432,
-            "MaintenanceDB": "${PGADMIN_CONFIG_DATABASE_NAME}",
-            "Username": "${PGADMIN_CONFIG_DATABASE_USERNAME}",
+            "MaintenanceDB": "${PGADMIN_DATABASE_NAME}",
+            "Username": "${PGADMIN_DATABASE_USERNAME}",
             "SSLMode": "prefer"
         }
     }
 }
 EOF
+fi
 
-echo "Bake CONFIG_DATABASE_URI in config_local.py:"
-export CONFIG_DATABASE_URI="postgresql://${PGADMIN_CONFIG_DATABASE_USERNAME}:${PGADMIN_CONFIG_DATABASE_PASSWORD}@${PGADMIN_CONFIG_DATABASE_HOST}:5432/${PGADMIN_CONFIG_DATABASE_NAME}"
+echo "Baking CONFIG_DATABASE_URI in config_local.py:"
 if [ ! -f /pgadmin4/config_local.py ]; then
     touch /pgadmin4/config_local.py
 fi
 echo "CONFIG_DATABASE_URI = '${CONFIG_DATABASE_URI}'" > /pgadmin4/config_local.py
 
-
-export PGADMIN_CONFIG_DATABASE_PASSWORD="\"$PGADMIN_CONFIG_DATABASE_PASSWORD\""
-export PGADMIN_CONFIG_DATABASE_USERNAME="\"$PGADMIN_CONFIG_DATABASE_USERNAME\""
-export PGADMIN_CONFIG_DATABASE_HOST="\"$PGADMIN_CONFIG_DATABASE_HOST\""
-export PGADMIN_CONFIG_DATABASE_NAME="\"$PGADMIN_CONFIG_DATABASE_NAME\""
+echo "Configuring pgAdmin, is done!"
+echo "Handing over to pgAdmin."
 
 /entrypoint.sh "$@"
